feat: upgrade to MobX v6
Upgrades to the library to support MobX 6. We had to wrap the task in an `action` to prevent MobX from automatically converting it to an action.

BREAKING CHANGE: The decorator will break when a MobX decorator is used on the same method. Additionally, updated the TypeScript target to es6, same as MobX 6.